functions/ingest-shard/requirements.txt
functions-framework>=3.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
google-genai
pydantic>=2.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
openpyxl>=3.1.0